+++ Roton.Interface/Video/Scenes/Composition/IBitmapSceneComposer.cs
@@ -7,8 +7,19 @@ namespace Roton.Interface.Video.Scenes.Composition
     /// </summary>
     public interface IBitmapSceneComposer : ISceneComposer, IDisposable
     {
+        /// <summary>
+        /// The underlying bitmap.
+        /// </summary>
         IDirectAccessBitmap DirectAccessBitmap { get; }
+
+        /// <summary>
+        /// If true, blinking characters will be replaced with a solid background color.
+        /// </summary>
         bool HideBlinkingCharacters { get; set; }
+
+        /// <summary>
+        /// If true, blinking is disabled and the full range of 16 colors can be used in the background.
+        /// </summary>
         bool UseFullBrightBackgrounds { get; set; }
     }
 }
